+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - [feat]: `setComponentSelf()` method;
 - [feat]: `set()` method;
 - [feat]: `normalizeSelf()` method;
+- [feat]: update input options for `add()`, `cross()`, `distance()`, `distanceSq()`, `dot()`, `equality()`, `reflect()`, `sub()` methods;
+- [breaking]: update `lerp()` default `coef` value to 1;
 
 ## 0.10.0 (2022-11-28)